refactor: extract server and window utilities
