+++ index.html
@@ -2209,6 +2209,7 @@
         });
     </script>
     
-    <script src="app.js"></script>
+    <script type="module" src="serverManager.js"></script>
+    <script type="module" src="app.js"></script>
 </body>
 </html>
